requirements.txt
 streamlit
 requests
 plotly
 google-genai
